requirements_fluid.txt
Flask
Flask-SocketIO
python-dotenv
eventlet
https://github.com/benoitc/gunicorn/archive/refs/heads/master.zip#egg=gunicorn==20.1.0
